requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
branca==0.8.1
bs4==0.0.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
distlib==0.3.9
filelock==3.18.0
folium==0.20.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
outcome==1.3.0.post0
pandas==2.3.0
platformdirs==4.3.8
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xyzservices==2025.4.0
uvicorn[standard]
fastapi
pymongo
python-dotenv
shapely